moving cluster.go to k88s
